requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.4
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extensions==3.1.3
django-phonenumber-field==5.2.0
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-generators==0.5.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy==0.902
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
phonenumbers==8.12.25
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0